+++ teku/src/main/java/tech/pegasys/teku/cli/options/BeaconRestApiOptions.java
@@ -104,6 +104,7 @@ public class BeaconRestApiOptions {
   @Option(
       names = {"--Xbeacon-liveness-tracking-enabled"},
       paramLabel = "<BOOLEAN>",
+      showDefaultValue = Visibility.ALWAYS,
       description = "Track validator liveness and enable requests to the liveness rest api.",
       arity = "0..1",
       fallbackValue = "true",